Move icons to own directory

+++ classy_finder.sh
@@ -1,21 +1,37 @@
 #!/bin/sh
+DEFAULT_FILENAME=classy_finder.png
 
 SCRIPT_DIR=`dirname $0`
 
-# read in image to use for finder icon
-DEFAULT_FILENAME=classy_finder@2x.png
-filename=${1:-$DEFAULT_FILENAME}
 
-# back up the originals to the local directory
-sudo cp /System/Library/CoreServices/Dock.app/Contents/Resources/finder.png ${SCRIPT_DIR}/finder_backup.png
-sudo cp /System/Library/CoreServices/Dock.app/Contents/Resources/finder@2x.png ${SCRIPT_DIR}/finder@2x_backup.png
+ICONS_DIR="${SCRIPT_DIR}"/Icons
+mkdir -p -- "$ICONS_DIR"
 
-# copy in the new ones
-sudo cp ${SCRIPT_DIR}/${filename} /System/Library/CoreServices/Dock.app/Contents/Resources/finder@2x.png
-sudo cp ${SCRIPT_DIR}/${filename} /System/Library/CoreServices/Dock.app/Contents/Resources/finder.png
+# read in image to use for finder icon (possibly from command line argument)
+filepath=${1:-${ICONS_DIR}/"$DEFAULT_FILENAME"}
 
-# trash the iconcache, forcing the dock to rebuild it
-sudo rm -f `find /private/var/folders -name com.apple.dock.iconcache 2> /dev/null`
 
-# restart the dock
-sudo killall Dock
+if [ -f "$filepath" ] ; then # image found
+	echo "New finder icon source: $filepath\n"
+
+	# back up the original to the local directory
+	sudo cp /System/Library/CoreServices/Dock.app/Contents/Resources/finder@2x.png ${ICONS_DIR}/finder_previous.png
+
+	# copy in the new one
+	sudo cp "$filepath" /System/Library/CoreServices/Dock.app/Contents/Resources/finder@2x.png
+	sudo cp "$filepath" /System/Library/CoreServices/Dock.app/Contents/Resources/finder.png
+
+	# trash the iconcache, forcing the dock to rebuild it
+	sudo rm -f `find /private/var/folders -name com.apple.dock.iconcache 2> /dev/null`
+
+	# restart the dock
+	sudo killall Dock
+
+	# output success
+	echo "Icon has been changed. The Dock will re-appear soon.\n"
+
+else
+	echo "No/invalid image selected."
+fi
+
+
